moved chat tests to test_sockets
